Did a sweep through items and blocks, along with slight tweaks to world
Changes: 
Items: Most of the items now should properly read from the config files on use
Blocks: Because of the flattening, I removed all methods related to meta-data and multiblocks. I decided to remove multiple of the Block classes and just register them as MoCBlocks to avoid cluttering up the workspace.
Dimensions: Registered the Wyvern Lair and created a DimensionType for it.
Rendering: I'm still pretty confused on how some of this works.